merge с родительским проектом

+++ application/models/Property.php
@@ -1,6 +1,7 @@
 <?php
 namespace app\models;
 
+use app\modules\shop\models\FilterSets;
 use app\properties\HasProperties;
 use app\properties\PropertyHandlers;
 use app\traits\GetImages;
@@ -10,6 +11,7 @@ use yii\caching\TagDependency;
 use yii\data\ActiveDataProvider;
 use yii\db\ActiveRecord;
 use yii\helpers\Json;
+use \devgroup\TagDependencyHelper\ActiveRecordHelper;
 
 
 /**
@@ -54,12 +56,12 @@ class Property extends ActiveRecord
             [
                 'class' => AttributeBehavior::className(),
                 'attributes' => [
-                    \yii\db\ActiveRecord::EVENT_BEFORE_INSERT => 'sort_order',
+                    ActiveRecord::EVENT_BEFORE_INSERT => 'sort_order',
                 ],
                 'value' => 0,
             ],
             [
-                'class' => \devgroup\TagDependencyHelper\ActiveRecordHelper::className(),
+                'class' => ActiveRecordHelper::className(),
             ],
             [
                 'class' => HasProperties::className(),
@@ -156,14 +158,12 @@ class Property extends ActiveRecord
     {
         /* @var $query \yii\db\ActiveQuery */
         $query = static::find()->where(['property_group_id' => $this->property_group_id]);
-        $dataProvider = new ActiveDataProvider(
-            [
-                'query' => $query,
-                'pagination' => [
-                    'pageSize' => 10,
-                ],
-            ]
-        );
+        $dataProvider = new ActiveDataProvider([
+            'query' => $query,
+            'pagination' => [
+                'pageSize' => 10,
+            ],
+        ]);
         if (!($this->load($params))) {
             return $dataProvider;
         }
@@ -213,13 +213,11 @@ class Property extends ActiveRecord
                     $cacheKey,
                     $prop,
                     0,
-                    new TagDependency(
-                        [
-                            'tags' => [
-                                \devgroup\TagDependencyHelper\ActiveRecordHelper::getObjectTag($prop, $id)
-                            ],
-                        ]
-                    )
+                    new TagDependency([
+                        'tags' => [
+                            ActiveRecordHelper::getObjectTag($prop, $id)
+                        ],
+                    ])
                 );
             }
             static::$identity_map[$id] = $prop;
@@ -229,7 +227,7 @@ class Property extends ActiveRecord
 
     /**
      * @param $group_id
-     * @return null|Property[]
+     * @return null|array<Property>
      */
     public static function getForGroupId($group_id)
     {
@@ -244,16 +242,14 @@ class Property extends ActiveRecord
                         $cacheKey,
                         $props,
                         0,
-                        new TagDependency(
-                            [
-                                'tags' => [
-                                    \devgroup\TagDependencyHelper\ActiveRecordHelper::getObjectTag(
-                                        PropertyGroup::className(),
-                                        $group_id
-                                    )
-                                ],
-                            ]
-                        )
+                        new TagDependency([
+                            'tags' => [
+                                ActiveRecordHelper::getObjectTag(
+                                    PropertyGroup::className(),
+                                    $group_id
+                                )
+                            ],
+                        ])
                     );
                 }
             }
@@ -292,7 +288,7 @@ class Property extends ActiveRecord
     }
 
     /**
-     *
+     * @inheritdoc
      */
     public function afterFind()
     {
@@ -321,6 +317,7 @@ class Property extends ActiveRecord
     }
 
     /**
+     * @inheritdoc
      * @param bool $insert
      * @return bool
      */
@@ -360,11 +357,11 @@ class Property extends ActiveRecord
         TagDependency::invalidate(
             Yii::$app->cache,
             [
-                \devgroup\TagDependencyHelper\ActiveRecordHelper::getObjectTag(
+                ActiveRecordHelper::getObjectTag(
                     PropertyGroup::className(),
                     $this->property_group_id
                 ),
-                \devgroup\TagDependencyHelper\ActiveRecordHelper::getObjectTag(Property::className(), $this->id)
+                ActiveRecordHelper::getObjectTag(Property::className(), $this->id)
             ]
         );
     }
@@ -396,22 +393,24 @@ class Property extends ActiveRecord
         foreach ($staticValues as $psv) {
             $psv->delete();
         }
-        if ($this->is_eav) {
-            $eavTable = $object->eav_table_name;
-            Yii::$app->db->createCommand()->delete(
-                $eavTable,
-                ['key' => $this->key, 'property_group_id' => $this->group->id]
-            )->execute();
+        if (null !== $object) {
+            if ($this->is_eav) {
+                Yii::$app->db->createCommand()->delete(
+                    $object->eav_table_name,
+                    ['key' => $this->key, 'property_group_id' => $this->group->id]
+                )->execute();
+            }
+            if ($this->is_column_type_stored) {
+                Yii::$app->db->createCommand()->dropColumn($object->column_properties_table_name, $this->key)->execute();
+                //                if ($object->object_class == Form::className()) {
+                //                    $submissionObject = Object::getForClass(Submission::className());
+                //                    Yii::$app->db->createCommand()
+                //                        ->dropColumn($submissionObject->column_properties_table_name, $this->key)
+                //                        ->execute();
+                //                }
+            }
         }
-        if ($this->is_column_type_stored) {
-            Yii::$app->db->createCommand()->dropColumn($object->column_properties_table_name, $this->key)->execute();
-            //                if ($object->object_class == Form::className()) {
-            //                    $submissionObject = Object::getForClass(Submission::className());
-            //                    Yii::$app->db->createCommand()
-            //                        ->dropColumn($submissionObject->column_properties_table_name, $this->key)
-            //                        ->execute();
-            //                }
-        }
+        FilterSets::deleteAll(['property_id' => $this->id]);
         parent::afterDelete();
     }
 
@@ -427,4 +426,3 @@ class Property extends ActiveRecord
         return null;
     }
 }
-
